# Conflicts:
#	drake/systems/controllers/pid_controller.cc
#	drake/systems/controllers/test/pid_controller_test.cc
#	drake/systems/framework/CMakeLists.txt
#	drake/systems/framework/primitives/CMakeLists.txt
#	drake/systems/framework/primitives/gain-inl.h
#	drake/systems/framework/primitives/gain.h
#	drake/systems/framework/primitives/test/CMakeLists.txt
#	drake/systems/framework/primitives/test/affine_linear_test.h
#	drake/systems/framework/primitives/test/affine_system_test.cc
#	drake/systems/framework/primitives/test/gain_test.cc
#	drake/systems/framework/primitives/test/linear_system_test.cc

+++ drake/systems/analysis/simulator.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <limits>
 #include <tuple>
 #include <utility>
 
@@ -336,7 +337,7 @@ void Simulator<T>::StepTo(const T& boundary_time) {
     const T next_update_dt = next_update_time - step_start_time;
 
     // TODO(edrumwri): Get the next publish time when API available.
-    T next_publish_dt = (double)std::numeric_limits<double>::infinity();
+    T next_publish_dt = std::numeric_limits<double>::infinity();
     T next_publish_time = step_start_time + next_publish_dt;
 
     // Attempt to integrate.
